+++ src/components/custom-field-input-number/custom-field-input-number.jsx
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React, { forwardRef, useImperativeHandle, useEffect, useRef, useState } from 'react';
-import styles from '../__internal__/abstract-custom-field.css';
+import styles from '../__internal__/abstract-custom-field/abstract-custom-field.css';
 import useValidation from '../../hooks/use-validation.jsx';
-import AbstractCustomField from '../__internal__/abstract-custom-field.jsx';
+import AbstractCustomField from '../__internal__/abstract-custom-field/abstract-custom-field.jsx';
 
 const apiLimits = {
   max: 2 ** 31,